// web/src/types/checkout.ts
export interface CheckoutFormData {
    customerName: string;
    customerPhone: string;
    city: string;
    district: string;
    neighborhood: string;
    fullAddress: string;
    deliveryTime: string;
    customTime: string;
    paymentMethod: string;
    cardNumber: string;
    cardName: string;
    cardExpiry: string;
    cardCvv: string;
    orderNote: string;
}